src/test/java/org/example/AlgorithmsRoadmap/ArraysAndHashing/ValidSudokuTest.java
package org.example.AlgorithmsRoadmap.ArraysAndHashing;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

public class ValidSudokuTest {

    @Test
    void isValidSudoku_validOne_true() {
        char[][] input = {
                {'5', '3', '.', '.', '7', '.', '.', '.', '.'},
                {'6', '.', '.', '1', '9', '5', '.', '.', '.'},
                {'.', '9', '8', '.', '.', '.', '.', '6', '.'},
                {'8', '.', '.', '.', '6', '.', '.', '.', '3'},
                {'4', '.', '.', '8', '.', '3', '.', '.', '1'},
                {'7', '.', '.', '.', '2', '.', '.', '.', '6'},
                {'.', '6', '.', '.', '.', '.', '2', '8', '.'},
                {'.', '.', '.', '4', '1', '9', '.', '.', '5'},
                {'.', '.', '.', '.', '8', '.', '.', '7', '9'},
        };

        assertTrue(ValidSudoku.isValidSudoku(input));
        assertTrue(ValidSudoku.isValidSudokuD(input));
    }

    @Test
    void isValidSudoku_duplicateInSubBox_false() {
        /*
        Explanation: Same as in previous, except with the 5 in
        the top left corner being modified to 9. Since there
        are two 8's in the top left 3x3 sub-box, it is invalid.
         */
        char[][] input = {
                {'9', '3', '.', '.', '7', '.', '.', '.', '.'},
                {'6', '.', '.', '1', '9', '5', '.', '.', '.'},
                {'.', '9', '8', '.', '.', '.', '.', '6', '.'},
                {'8', '.', '.', '.', '6', '.', '.', '.', '3'},
                {'4', '.', '.', '8', '.', '3', '.', '.', '1'},
                {'7', '.', '.', '.', '2', '.', '.', '.', '6'},
                {'.', '6', '.', '.', '.', '.', '2', '8', '.'},
                {'.', '.', '.', '4', '1', '9', '.', '.', '5'},
                {'.', '.', '.', '.', '8', '.', '.', '7', '9'},
        };

        assertFalse(ValidSudoku.isValidSudoku(input));
        assertFalse(ValidSudoku.isValidSudokuD(input));
    }

    @Test
    void isValidSudoku_duplicateInColumn_false() {
        char[][] input = {
                {'5', '3', '.', '.', '7', '.', '.', '.', '.'},
                {'6', '.', '.', '1', '9', '5', '.', '.', '.'},
                {'.', '9', '8', '.', '.', '.', '.', '6', '.'},
                {'8', '.', '.', '.', '6', '.', '.', '.', '3'},
                {'4', '.', '.', '8', '.', '3', '.', '.', '1'},
                {'7', '.', '.', '.', '2', '.', '.', '.', '6'},
                {'.', '6', '.', '.', '.', '.', '2', '8', '.'},
                {'.', '.', '.', '4', '1', '9', '.', '.', '5'},
                {'4', '.', '.', '.', '8', '.', '.', '7', '9'},
        };

        assertFalse(ValidSudoku.isValidSudoku(input));
        assertFalse(ValidSudoku.isValidSudokuD(input));
    }

    @Test
    void isValidSudoku_duplicateInRow_false() {
        char[][] input = {
                {'5', '3', '.', '.', '7', '.', '.', '5', '.'},
                {'6', '.', '.', '1', '9', '5', '.', '.', '.'},
                {'.', '9', '8', '.', '.', '.', '.', '6', '.'},
                {'8', '.', '.', '.', '6', '.', '.', '.', '3'},
                {'4', '.', '.', '8', '.', '3', '.', '.', '1'},
                {'7', '.', '.', '.', '2', '.', '.', '.', '6'},
                {'.', '6', '.', '.', '.', '.', '2', '8', '.'},
                {'.', '.', '.', '4', '1', '9', '.', '.', '5'},
                {'.', '.', '.', '.', '8', '.', '.', '7', '9'},
        };

        assertFalse(ValidSudoku.isValidSudoku(input));
        assertFalse(ValidSudoku.isValidSudokuD(input));
    }
}
